src/requirements.txt
python-telegram-bot==20.7
sqlalchemy>=2.0.25
aiosqlite==0.19.0
telethon==1.34.0
cryptography==41.0.7
